refactored code using context api.
