+++ react-router-project/src/Header.js
@@ -1,6 +1,9 @@
 import { FaLaptop, FaTabletAlt, FaMobileAlt } from "react-icons/fa";
+import useWindowSize from "./hooks/useWindowSize";
 
-const Header = ({ title, width }) => {
+const Header = ({ title }) => {
+  const { width } = useWindowSize();
+
   return (
     <header className="Header">
       <h1>{title}</h1>
