+++ react-router-project/src/Home.js
@@ -1,6 +1,10 @@
 import Feed from "./Feed";
+import { useContext } from "react";
+import DataContext from "./context/DataContext";
 
-const Home = ({ posts, fetchError, isLoading }) => {
+const Home = () => {
+  const { searchResults, fetchError, isLoading } = useContext(DataContext);
+
   return (
     <main className="Home">
       {isLoading && <p className="statusMsg">Loading Posts...</p>}
@@ -11,8 +15,8 @@ const Home = ({ posts, fetchError, isLoading }) => {
       )}
       {!isLoading &&
         !fetchError &&
-        (posts.length ? (
-          <Feed posts={posts} />
+        (searchResults.length ? (
+          <Feed posts={searchResults} />
         ) : (
           <p className="statusMsg">No Posts to Display</p>
         ))}
