+++ react-router-project/src/Layout.js
@@ -2,14 +2,17 @@ import Header from "./Header";
 import Nav from "./Nav";
 import Footer from "./Footer";
 import { Outlet } from "react-router-dom";
+import { DataProvider } from "./context/DataContext";
 
-const Layout = ({ search, setSearch, width }) => {
+const Layout = () => {
   // need to pass in search and setSearch for the Nav component
   return (
     <div className="App">
-      <Header title="React JS Blog" width={width} />
-      <Nav search={search} setSearch={setSearch} />
-      <Outlet />
+      <Header title="React JS Blog" />
+      <DataProvider>
+        <Nav />
+        <Outlet />
+      </DataProvider>
       <Footer />
     </div>
   );
