+++ react-router-project/src/context/DataContext.js
@@ -1,0 +1,68 @@
+import { createContext, useState, useEffect } from "react";
+import useAxiosFetch from "../hooks/useAxiosFetch";
+
+const DataContext = createContext({});
+
+export const DataProvider = ({ children }) => {
+  const [posts, setPosts] = useState([]);
+  const [search, setSearch] = useState("");
+  const [searchResults, setSearchResults] = useState([]);
+  const { data, fetchError, isLoading } = useAxiosFetch(
+    "http://localhost:5500/posts"
+  );
+
+  useEffect(() => {
+    setPosts(data);
+  }, [data]);
+
+  // useEffect(() => {
+  //   const fetchPosts = async () => {
+  //     try {
+  //       const response = await api.get("/posts");
+  //       // axios automatically converts to JSON and catch errors when they arent't in the 200 range
+  //       // this means we don't have to do an if condition to make sure we have a response
+  //       setPosts(response.data);
+  //     } catch (err) {
+  //       if (err.response) {
+  //         // not in the 200 response range
+  //         console.log(err.response.data);
+  //         console.log(err.response.status);
+  //         console.log(err.response.headers);
+  //       } else {
+  //         console.log(`Error: ${err.message}`);
+  //       }
+  //     }
+  //   };
+
+  //   fetchPosts();
+  // }, []);
+
+  useEffect(() => {
+    const filteredResults = posts.filter(
+      (post) =>
+        // searching the body or title to see if it includes the search characters
+        post.body.toLowerCase().includes(search.toLowerCase()) ||
+        post.title.toLowerCase().includes(search.toLowerCase())
+    );
+
+    setSearchResults(filteredResults.reverse());
+  }, [posts, search]);
+
+  return (
+    <DataContext.Provider
+      value={{
+        search,
+        setSearch,
+        searchResults,
+        fetchError,
+        isLoading,
+        posts,
+        setPosts,
+      }}
+    >
+      {children}
+    </DataContext.Provider>
+  );
+};
+
+export default DataContext;
